- made possible classes sealed for performance - organized different sections for tabs in UI - changed cache system in textures
