+++ netware/Configuration/Config.cs
@@ -8,7 +8,6 @@ using NetWare.Utils;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Reflection;
