+++ netware/Modules/Data.cs
@@ -4,4 +4,5 @@ public static class Data
 {
     public static bool ResetFov { get; set; }
     public static bool ResetZoom { get; set; }
+    public static string SelectedConfig { get; set; }
 }
