+++ netware/NetWare.csproj
@@ -271,6 +271,18 @@
     <Compile Include="Models\Range.cs" />
     <Compile Include="Modules\MenuTabs\IMenuTab.cs" />
     <Compile Include="Modules\MenuTabs\MenuTab.cs" />
+    <Compile Include="Modules\Tabs\CombatTab\AimbotSection.cs" />
+    <Compile Include="Modules\Tabs\CombatTab\SilentAimSection.cs" />
+    <Compile Include="Modules\Tabs\SettingsTab\ConfigManagerSection.cs" />
+    <Compile Include="Modules\Tabs\SettingsTab\ConfigSelectionSection.cs" />
+    <Compile Include="Modules\Tabs\SettingsTab\GameplaySection.cs" />
+    <Compile Include="Modules\Tabs\SettingsTab\WatermarkSection.cs" />
+    <Compile Include="Modules\Tabs\VisualTab\BoxesSection.cs" />
+    <Compile Include="Modules\Tabs\VisualTab\CameraSettingsSection.cs" />
+    <Compile Include="Modules\Tabs\VisualTab\FovChangerSection.cs" />
+    <Compile Include="Modules\Tabs\VisualTab\NameTagsSection.cs" />
+    <Compile Include="Modules\Tabs\VisualTab\SkeletonSection.cs" />
+    <Compile Include="Modules\Tabs\VisualTab\TracersSection.cs" />
     <Compile Include="Natives\Shell32Lib.cs" />
     <Compile Include="UI\MenuChecks.cs" />
     <Compile Include="UI\Menu.cs" />
